Version 2.2.0
Support for tolerance amounts has been added, meaning that if an
order wasn't fully paid but Talo reported it as successful because
the difference is within that margin, we  create an invoice for the
difference with a related note.

This allows the order to change status later if the shipment has
been created (preventing orders from being blocked due to
non-payment of the original total).